Don't add throw-away DocumentMeta to test DB
`factories.Annotation()` creates a `DocumentMeta` object that isn't
meant to be added to the DB - it's created using
`factories.DocumentMeta.build()`, as opposed to just
`factories.DocumentMeta()`, which means "build the object but don't add
it to the DB session" (see the factory_boy docs).

`factories.Annotation` just wants to make use of
`factories.DocumentMeta`'s random value generation: it creates a dict
using values from the `DocumentMeta` object but discards the actual
`DocumentMeta` itself.

Further down in `factories.Annotation` these document meta dicts end up
getting passed to `update_document_metadata()`.
`update_document_metadata()` will create _another_ new `DocumentMeta`
based on the values in the dict **unless** a duplicate `DocumentMeta`
already exists in the DB, then it will reuse the existing one instead.
This avoids any `IntegrityError`s due to adding duplicate
`DocumentMeta`s to the DB.

The **problem** is that `factories.Annotation` sets the `document`
attribute of the throw-away `DocumentMeta` to `self.document`.
`self.document` is an ORM object that _is_ added to the DB session, so
this means that the throw-away `DocumentMeta` ends up indirectly getting
added to the DB session as well.

Since there's nothing to prevent the throw-away `DocumentMeta` from
being a duplicate of one that's already in the DB, and
`factories.DocumentMeta`'s random generation will sometimes produce
duplicates, this means that tests that use `factories.Annotation()` will
sometimes fail with `IntegrityError`.

There seems to be no need to set the `document` attribute, so fix the
problem by simply not setting it.

`factories.DocumentMeta()` would normally generate a random new document
to be the `DocumentMeta`'s document. We don't want this either so
explicitly pass `document=None`.