Fix some failing tests
These tests create a `Document` and then an annotation of that document,
for example:

    document = factories.Document()
    results = [factories.Annotation(document=document,
                                    updated=annotation_datetime)]

and the rest of the test implicitly assumes that
`document == results[0].document`, but this assumption doesn't always
hold true.

Because of the way the document equivalence system in Hypothesis
(`Document`, `DocumentURI`, `DocumentMeta` and related functions) works,
creating a new `Annotation` of a given `Document` can sometimes cause a
"document merge" to happen in which different `Document` objects in the
DB are deemed to be equivalent. All but one of these duplicated
`Document`s is removed from the DB and `annotation.document` is changed
to point to the one remaining `Document`.

In particular `factories.Annotation()` can cause these merges to happen.

All of this means that this assertion will usually work but sometimes fail:

    document   = factories.Document()
    annotation = factories.Annotation(document=document)
    assert document == annotation.document

Fix the tests to remove the assumption that `document == annotation.document`
and just use `annotation.document` directly instead.