Share subtitle count class for subtitle counts
Previously there were two css classes that did the same thing,
this change simply removes one of them and changes the template
to use the same css class as the other subtitle for formating
the number that comes after the subtitle. It also moves the font
size into a variable located in the _base.