Add tests for logo and authority in group & search
Add tests for logo property in group. If the logo is in the dict
of logos it should return the logo, if not it should return None.
Add tests for logo and authority in group search results. If the
authority is the default authority it should be None, otherwise
it should return the group's authority. If the logo is in the dict
of logos it should return the logo, if not it should return None.