Fix flakey factories.User
factories.User was generating usernames and email addresses that were
not guaranteed to be unique. This meant that any test that called
factories.User more than once would sometimes get two users with the
same username and/or email, which will result in an IntegrityError from
the DB if something adds these users to the DB session then tries to flush it.
The end result is flakey tests that sometimes fail.

The email address now also includes the username in it, instead of a
different random name.

Fixes:

* https://github.com/hypothesis/h/issues/4851
* https://github.com/hypothesis/h/issues/4853
* https://github.com/hypothesis/h/issues/4854
* https://github.com/hypothesis/h/issues/4855
* https://github.com/hypothesis/h/issues/4856
* https://github.com/hypothesis/h/issues/4858
* https://github.com/hypothesis/h/issues/4859

Before:

    > u = factories.User()
    > u.username
    'cbruce'
    > u.email
    'karen@king-mcmahon.biz'

After:

    > u = factories.User()
    > u.username
    'llucas__0'
    > u.email
    'llucas__0@gmail.com'

(The first username generated will end in "0", the next call to
factories.User() will return a user with a username like "<SOMEONE>__1"
and so on. The <SOMEONE> part is randomly generated and then the number
appended to the end guarantees uniqueness by always counting upwards.)