Remove error state of form inputs after editing them
Remove the bold red border and validation error messages after editing a
form input, since the validation error message is no longer applicable
at that point.

Form inputs were already being rendered with a `js-form-input` class.
Add a controller for those elements which removes the "is-error" class
after the child `<input>` is edited.

Note that the `is-error` class controls both the border style and
visibility of validation error messages.