Import and translate sequence and sequence_item templates from Deform
As a starting point for supporting editing lists of things in our forms,
import the "sequence.pt" and "sequence_item.pt" templates for sequence
fields from the deform package and roughly translate them to Jinja 2.

It doesn't work yet because the templates still reference non-existent
JavaScript and violate our CSP rules, and looks ugly because there is no
styling.