Initial implementation of a form control for editing lists
Implement a form control for editing form fields that have a list type.
These fields use the `SequenceSchema` colander schema and deform's
`SequenceWidget`.

I started with deform's bundled templates for `SequenceWidget` and
rewrote them to conform to our conventions for form controls:

 - BEM-style CSS
 - `<template>` elements for item templates that need to be "stamped
   out" on the client side
 - Use of correct semantic elements for things (eg. `<button>` for
   buttons)
 - Removal of inline JavaScript (which our CSP policy disallows) and
   logic added to the HTML via a `Controller` subclass.

The other form controls were designed to work without JavaScript. I have
intentionally not done that here as I don't think we have a need.

This initial implementation is not pretty and it doesn't implement some
of the features of the original sequence widget, such as item
re-ordering, but it does provide a basic scaffold that we can build on
in future.