Add organization model
Organization is a group of groups. Currently this is a very simple
implementation with attributes logo, name, pubid, and authority.
The name is the name of the orgnaization, the logo is expected to
be an svg, the pubid is a unique public identifier to be used in
urls, and authority is the authority of the group. A cap was put
on the size of the logo in chars simply to prevent a malliciously
large upload. It also verifies the logo is an svg and is
htmlparseable.