Add associated_groups to list_groups service
Associated groups are groups that a particular user either is a
member of or a creator of. This list of groups includes all group
types. Associated groups will be used to populate Your groups in
the group search facet on the search page. This also changes
the open_groups method to readable_by_world_groups as this is no
longer a list of open groups since restricted groups were added.
Also got rid of the deprication message as this method will need
to be kept for associated_groups even after the removal of
all_groups. Added testing for associated groups and adjusted
fixtures and one test expected value due to fixture changes.