Add Creator label to My group facet
Add a relationship label after the title of the group on the search
autocomplete group facet drop down. Currently the only relationship
so far is 'Creator' but there may be more relationships defined
later. Note this also changes the name Your groups to My groups.
A styling class for the new relationship label was added. The navbar
tests were also updated to reflect the relationship addition in the
suggested_groups.