Add metric to record when badge count is zero
This uses new relic's custom metrics to record when the badge count
is zero. If the badge count is zero it returns 1, otherwise it
returns zero. Because we record it everytime, we can easily get a
percentage of the transactions that are zero. The average value,
will give us a value between 0-1 that is the percentage of times
the badge count is zero.