Add metric unAuthUserGotZero & event BadgeNotZero
UnAuthUserGotZero is, out of all the requests that were zero,
the percentage of them where the user was None/unauthenticated.
This will tell us if a large portion of the badge returning
zero requests are due to unauthenticated users.

BadgeNotZero is a count or the number of requests that the
given user got a badge count that wasn't zero. This is an event
meaning it will be query-able and we can combine data from all
users onto the same chart. This will tell us if there are a small
portion of users for which the badge count is consistently
reporting non zero.