Remove model-level validation of SVGs
Since organizations and their logos can only be created by developers
currently I don't think this is needed, and this validation currently
prevents valid SVG values from being set: many SVG files start with an
`<?xml>` tag not an `<svg>` tag.

If we do have a form for uploading SVGs in the future I'm not sure that
the model will be the right place to validate them. For example we might
want to have a separate validation function that's called by the view
before the SVG is ever passed on to the service or model layers.