Don't crash when logged out user tries to leave group
1. Login and create or join a group.
2. Open the group's page in a tab.
3. Open Hypothesis in another tab and log out.
4. Back in the first tab, click the leave group button.
=> You'll see a 500 Server Error page.

The code involved in leaving a group doesn't assume that the user is a
member of the group (if the user isn't a member, e.g. they already left
the group, then clicking the leave group button again is a no-op).

But it does assume that a user is logged in, and if one isn't it ends up
with a TypeError in the groups service.

Fix this by turning it into a 404 page instead.

Fixes https://github.com/hypothesis/h/issues/4917.