Add ability to search for groups by name in admin form
In h instances with a large number of groups it becomes impractical to
find a specific group by paging through the list, so add a facility to
search for groups by name.

The search currently does a naive `LIKE` SQL query which should be OK
for the admin page. If we were to expose this to users we might need to
optimize it using a suitable index.