Add admin CRUD views for managing organizations
Add views for:

 - Listing organizations
 - Creating and editing organizations
 - Deleting organizations

Since our form library does not have a file upload widget, the logo
input is just a text area where the user is expected to paste the SVG
content of the logo.

Organizations can only be deleted if they are not associated with any
groups.

Fixes https://github.com/hypothesis/product-backlog/issues/522