Create world group in default organization
When the DB is not currently (or yet) being managed by Alembic
`h/db/__init__.py` creates the `__world__` group and the `__default__`
organization when h starts up.

Change this code to set the `__world__` group's `organization_id` to
that of the `__default__` organization. This will be needed in the
future when `group.organization_id` becomes not nullable, otherwise
creating the `__world__` group would crash.