Use default organization in group model tests
Once we've made `group.organization_id` not nullable in the database,
`models.Group` unit tests that try to add a group with no organization
to the DB and flush it will crash.

Fix these tests to add the default organization to the groups.

Tests that don't add the `Group` objects to the DB or flush it don't
need to be updated.

One test, `test_group_organization_defaults_to_none()`, is deleted
because it's no longer possible to add a group with no organization to
the DB and flush it (as this test does) without crashing.
`Group.organization` still defaults to `None` _before_ you flush the DB,
but this is just the standard behaviour of an attribute of a sqlalchemy
ORM class (or any Python class) if you don't set it to a value, so it
doesn't need to be tested.