Implement new admin navbar design
The existing navbar covered the main content of the page when it wraps
onto multiple lines, and the wrapping happens even on not-very-narrow
screens. This commit implements a revised look for the navbar in the
admin panel which resolves that problem and generally uses horizontal
space more efficiently.

As a bonus, the layout is implemented with flexbox following our CSS
conventions which should make it easier to adapt in future.