Allow users to set organization when creating or editing a group
Add a dropdown list of organizations to the admin group create/edit form
which allows the user to select the org that the group will belong to.
This defaults to the default organization.

Since organizations are associated with an authority, it is not
necessary for the user to separately specify an authority when creating
a group. Instead this can be inferred from the organization.

When editing an existing group, the user can only re-assign the group to
another organization belonging to the same authority.