Run prospector as well as flake8 on Travis
I don't want to replace flake8 with prospector in `make lint` yet
because of the amount of stuff that has to be ignored or disabled in
prospector to get it passing.

But I don't want new prospector errors to creep in while we're trying to
fix them all either.

So add prospector as an additional step on Travis.