Correct alt text
Correct an HTML `alt` text attribute in an `<img>` element from "Org:"
to "Organization logo".

The alt text it just supposed to identify the picture. It is used by
assistive technology (e.g. screen readers) to describe the image to the
user. It's also used by search engines to try to understand what the
content of the image is, and some image search engines will display the
alt text next to the image in search results. So "Org:" won't work.

It's true that browsers will also display the alt text in place of the
image if for example the user is browsing the web with images turned
off or if the image fails to load. But this situation is much rarer than
screen readers and search engines, and such users are used to seeing
descriptions / identifications of the images in place of the images,
like "Organization logo", rather than text that's intended to be part of
the page layout like "Org:".