Remove creator and members from results['group']
The creator and members keys are no longer used by the template.
They were moved outside of the group dict to the group_user_args
list. The second value in the list is the members and the third
value is the group creator. When this move occured it looks like
the dev forgot to remove the members and the creator from the
group dict. Since they were never removed, the change in the tests
was never made to use the new group_user_args. This commit fixes
that.