Move _group_principals helper into class
Move the private helper function `_group_principals` into the
`AnnotationResource` class, making it a `staticmethod` since it doesn't need
access either `self` or `cls`.

This is just a teeny bit of code organization / cleanup: since this
function is only used in this class, define it in the class where it is
used, rather than far away at the bottom of the module.

I'm not a big fan of the long disorganised list of helper functions at
the bottom of the file design pattern, I don't think it helps code
organisation or readability.