Cleanup GroupSearchController tests
This cleans up the GroupSearchController tests. A lot of these tests
were implemented incorrectly in the way they handled permissions.
This mocks the has_permissions on the request object to behave as
expected with the various group and user combinations. The tests were
also simplified (aka where all groups share the same logic in the
GroupSearchController class, only a private group use case is tested).
Note, one permissions test was removed entirely; the test that asserts
a raises HttpNotFound when the group doesn’t exist as this is not in
sync with the current behavior. The HttpNotFound is raised before ever
reaching this class in the app.