Change Client attributes into @property's
Change h.search.client.Client instance variables to @property's.

While this is a bit more verbose, it means that index and conn are now
visible to the mock library's create_autospec() function (mock Client
instances created using create_autospec() now have index and conn
attributes, which they wouldn't before).

See:

https://docs.python.org/3/library/unittest.mock.html#autospeccing

the paragraph beginning "A more serious problem is that it is common for
instance attributes to be created in the __init__() method and not to
exist on the class..."

This change also clarifies the API of the class: makes it explicit that
index and conn are read-only.