Pass ELASTICSEARCH_HOST env var to unit tests
Tell tox to pass the ELASTICSEARCH_HOST environment variable through to
the code running in the test environment.

tox.ini is already configured to do this for the functional tests, but
not for the unit tests.

Since we're going to be adding unit tests that connect to Elasticsearch
the unit test environment is gonna need ELASTICSEARCH_HOST too.