Add Elasticsearch 6.2 to Jenkins
Run an Elasticsearch 6.2 Docker container on Jenkins, using the same
Elasticsearch 6.2 docker image as we use in development environments, in
addition to the existing Elasticsearch 1.6 Docker container that already
runs on Jenkins.

Rename the existing `elasticsearch` to `elasticsearch_old` and
`elasticsearchHost` to `elasticsearchHost_old`. Then add new
`elasticsearch` and `elasticsearchHost` for the new, Es 6.2 docker
container.

Also pass the URL of this new container through to the tests using the
new `ELASTICSEARCH_URL` env var.

Also add the new `ELASTICSEARCH_URL` env var to `tox.ini` so that it
passes it through to our code.