Rename get_client to _old and add new get_client
Make get_client connect to the elasticsearch6.2 cluster and rename
the original get_client that connects to the elasticsearch1.5 cluster
to get_client_old. Note the new get_client does not return a custom
wrappered client class object but simply the es client obj directly.
The Client class was refactored to pass the extra config settings
directly to the es client class without adding any additional config.
A _get_settings function was added since most of the settings are shared
between the old and new clusters. es.url will be the es.host for the
6.2 cluster. This name was chosen because it is more explicit than
the original es.host. Note the elasticsearch client object (aka es) is
no longer attached to the request but rather set as the default connection
to use in elasticsearch_dsl. This is the recommended way of specifying
the connection and it means we don't need to pass in using='es' everywhere
in our elasticsearch_dsl code.