Make search_replies() a private method
search_replies() is a private method - it's called by the run()
method in the same class, and all users of this class call run() never
search_replies().

So rename it to _search_replies() for clarity.

This requires updating lots of tests that are calling
search_replies(). I think this just reveals and existing problem
with the code design / tests.