Make default_querybuilder() a private method
default_querybuilder() is a private method of the Search class: it's
called only by methods of Search and never called from anywhere else.

Clarify this by renaming it to _default_querybuilder(). Also make it a
@staticmethod within the class itself, just to clarify that it belongs
to this class.

This requires updating some tests that're calling default_querybuilder()
directly. I think this just reveals an existing problem with the code
design and tests.