Make search_annotations() a private method
search_annotations() is a private method - it's called by the run()
method in the same class, and all users of this class call run() never
search_annotations().

So rename it to _search_annotations() for clarity.

This requires updating lots of tests that are calling
search_annotations(). I think this just reveals and existing problem
with the code design / tests.