Add Es-integrated tests for separate_replies
Add the first Elasticsearch-integrated tests for the search querying
functionality.

These test elements of the search querying functionality related to the
separate_replies argument: how search behaves with and without this
argument.

I'm not sure that everything to do with separate_replies is covered by
these tests yet. For example deleted replies are excluded from
reply_ids, and this isn't tested yet. There are several more things like
this still to test.