Generate test annotation ids in factory
Annotations created with factories.Annotation() do have ids but
unsaved and stub annotations created by factories.Annotion.build() or
factories.Annotation.stub() without passing an id argument don't have
ids:

    >>> a = factories.Annotation.build()
    >>> a.id
    None

This is because the ids are generated by models.Annotation using a
server_default function, so an id isn't generated until the annotation
is actually saved to the DB.

This creates noise in tests that have to pass an id argument to build()
or stub() because, for example, the test is going to add the annotation
to the test Elasticsearch index and doing so requires a non-null id,

Add code to factories.Annotation to auto-generate an id when build() or
stub() is called without passing an id.

Annotations created without using build() or stub() are not affected,
they still get an id from the server_default function:

    >>> factories.Annotation()
    <Annotation 3nemKFqbEeirh6fGyzgKag>

Annotations created with build() or stub() now also get a UUID id.
You'll notice that these look slightly different to the server_default
ids because the code for creating the ids isn't exactly the same. But
it's good enough for the tests:

    >>> factories.Annotation.build()
    <Annotation MTVhOTNiZDQtNWE5Yy0xMW>

    >>> factories.Annotation.stub().id
    'MjFlOTMwZmMtNWE5Yy0xMW'

You can still pass an id argument to create an annotation with a
particular id:

    >>> factories.Annotation.build(id="foo")
    <Annotation foo>

    >>> factories.Annotation.stub(id="bar").id
    'bar'