Add SettingsManager to replace EnvSettings
SettingsManager accepts the pyramid config settings dict and the
environ dict on init and has a set method that is used to set a
setting in the pyramid config settings dict. The setting may be
left untouched, it may be set to the environment variable, or
it may be overriden/set to the default setting provided. Things
like logging a deprecation message, erroring if the setting is
required and is not set by the time set() returns, and type
casting the setting are also included in the set() method.