Add request.user to pyramid_request fixture
h adds a custom `.user` property to the request but Pyramid's
DummyRequest (which is what we use for our pyramid_request fixture)
doesn't have this property. As a result lots of tests need to add
`pyramid_request.user` so that code doesn't crash.

Add `pyramid_request.user` by default so tests no longer need to do
this.