Remove delete_all_elasticsearch_documents fixture
The es_client fixture is responsible for creating the elastic search
client connection and cleaning it up after it is used in the test so
the special delete_all_elasticsearch_documents fixture is no longer
needed. The always_delete_all_elasticsearch_documents fixture that is
called in every functional test was left but it now simply calls
the es_client fixture underneith. It's possible that the
always_delete_all_elasticsearch_documents could go away as most likely
the only place where a test is modifying es is where the es_client is
used (at least one would hope) but for now we'll keep it.