Replace HKDF implementation from cryptography
The `cryptography` Python library currently fails to build against the
latest version of `libressl-dev` in the Alpine release (3.7) that we are using.
`libressl-dev` is in turn a dependency of `postgresql-dev`.

Fortunately we only use `cryptography` for the well-known HKDF key
derivation function. This commit replaces the `cryptography` package
with the very small `hkdf` package which only depends on `hashlib` and
`hmac` from the Python standard library.

Add tests for the output of `derive_key` based on the vectors from the
HKDF RFC to verify that the outputs are the same as with the previous
implementation.