Pass a `List[str]` instead of `str` as `hosts` argument
The `hosts` argument to `Elasticsearch` can in fact be a single string
rather than a list of strings or dicts, but this is not a documented
feature. Pass a list containing a single string since that is a
documented way to use the API and also what we do in the tests.