Mark `SECRET_KEY` as a required setting
The `SECRET_KEY` value is used to derive the key for signing cookie
sessions. The previous behavior of generating a random value if this is
not set is not useful in a production environment because each web
process generates a different value and therefore cookies signed by one
process fail to validate when submitted to a different web process. Not
setting a `SECRET_KEY` also means that existing cookies become invalid
after a server restart.