Make shared search fixture w/ no default filters
Make a module-level search fixture that is used by each individual
filter test class that has no default filters defined. Thus we can
be sure we are only testing one filter at a time. The class fixtures
were renamed to filtered_search so that the generic search fixture
can be easily identified as seperate from the class filtered_search
fixture that applies a filter to the search.