Improve a code comment
Don't use the word "inherits" because object oriented inheritance isn't
what's going on here.

Do use "default root factory" and "lineage" as these are the Pyramid
terms (to be found in the Pyramid docs) that're relevant here.

Also explain *why* we're adding the default root factory and
specifically mention the "admin_oauthclient" permission so that it'll
show up when you grep for that.

Also add a link to the relevant docs.