Add a string representation to UserIdentity
This is used for example by the Python shell when it renders a
UserIdentity:

    >>> UserIdentity(provider=u"lms", provider_unique_id=u"abc123")
    UserIdentity(provider=u'lms', provider_unique_id=u'abc123')

I've designed the repr string so that if you were to copy-paste it then
it would be valid Python for constructing a new UserIdentity the same as
this one. I think this is a good general approach to repr strings (when
it can be done easily).