Fix email validator to allow null emails
Fix the model-level email validator to allow for None as a valid email
address.

Since User.email can be None now, validate_email() needs to allow None
as a valid value for an email.

Previously it was possible to create a user without an email address by
simply not specifying an email, as it would default to null:

    user = User(username="seanh", authority="hypothes.is")

But if you tried to explicitly specify None as the email value you'd get
an error because validate_email() assumes the email is a string and
crashes:

    user = User(..., email=None)
    TypeError: object of type 'NoneType' has no len()

Similarly if you took an existing user with a non-null email address and
did `user.email = None`, you'd get the same crash.

Fix this crash by returning None emails as valid values.