Handle null emails in get_by_email()
Now that User.email can be None, get_by_email() needs to handle users
with no emails.

It _could_ just assume that its `email` argument is never None, but
given that User.email can be None now and code might be doing
get_by_email(..., user.email, ...) with some User object, it's probably
a good idea for get_by_email() to be able to handle the `email` argument
being None.

There could be many users in the database with None for their email, and
get_by_email() has so far always returned a single matching user or
None, and returning a list of all users that don't have an email isn't
very useful, so when the given `email` is None just return None.