Add a get_es6_client and re-use the`Client`class
Add a parameter called elasticsearch which is the elasticsearch library to use to
create the elastic search connection in the Client class

Move all shared kwargs from inside the Client class and from inside the get_client
method into a seperate method specificially for generating the shared connection
settings.

Add a get_es6_client method that gets the Elasticsearch 6 connection and expose it
on the request as `request.es6`.

This will enable existing code for setting up the Elasticsearch index
and reindexing annotations to be used with the ES6 cluster.