Add annotation mapping for Elasticsearch 6
Add mappings that define the schema of indexed annotations for
Elasticsearch 6. Compared to the existing `ANNOTATION_MAPPING`, the
following changes have been made:

 - The indexed fields are sorted alphabetically
 - "string" fields have been changed to use either "text" or "keyword"
   types depending on whether we need to support search within those
   fields ("text") or only on the whole content of the field ("keyword")
 - Fields which we have no need to search or sort on have been removed.
   This includes:

    - `annotator_schema_version`
    - `target.selector` fields other than `exact` (the annotation's
      quote)
 - The "id" field has been explicitly added since the document's `_id`
   field is not configurable in ES6.
 - The `dynamic` option has been set to `false` to disable indexing of
   any fields which are not explicitly listed in the mapping. This
   avoids anyone relying on being able to search fields which happen to
   get written to the index but are not explicitly listed as searchable.

The analysis settings have not been changed since the existing settings
work with ES6.

Since the mapping depends on which es client is passed in (es1 or es6)
the mapping was exposed as a parameter to the higher level code as the
higher level code knows which version of the client it's passing.

Catch both es1 and es6 exceptions.

Add tests for updated_index_settings to verify that both es1 and es6
errors are handled correctly.

Parametrize test for get_aliased_index to verify that es6's errors
are handled correctly.

Select mapping based on new es version client property.