Index annotations into ES6 when feature flag is on
When the `index_es6` feature flag is enabled, index annotations into
both Elasticsearch 1 and Elasticsearch 6.

During reindexing, annotations need to be indexed into the new index.
Since it is possible we may be rebuilding only one of the indexes (ES1
or ES6), a separate setting has been added in the DB to record the name
of the new ES6 index.