Remove SETTING_NEW_INDEX constant
While this constant is used in multiple places, it doesn't provide
any more context to the string itself, it is unlikely to change,
and having to import it everywhere adds a layer of uneccesary
indirection. Thus the constant was removed and where ever it was used
was replaced with the value itself.