Make `/api/badge` requests cheaper for URIs that have never been annotated
For each request to the badge endpoint, the view function:

 1. Looks up the URI in a blocklist in the DB
 2. Performs a search for annotations against a given URI which involves:
    1. Expanding the URI into all "equivalent" URIs via a DB query
    2. Enumerating groups that the user has read access to
    3. Issuing a query to ES with the complete query

Data suggests that the majority of URIs passed in queries to the badge
endpoint have never been annotated. Therefore perform a cheap index-only
DB query against the `document_uri.uri_normalized` field to check if the
URI has ever been annotated and skip the rest of the request if this
returns false.

There is still a fair amount of per-request overhead not eliminated by
this approach, but this is easy to implement and should still be a
significant win.