Fix flakey AnyMatcher tests
Any matcher matches against several different fields inside of an
annotation. The test factory randomly sets fields that aren't
explicitly set and sometimes these settings conflict with the test.
Aka if we search for any="bar", the text field may be randomly set
to a string that includes bar therefore we need to set these fields
explicitly in the test and not depend on the random setting from
the factory.