Optimize redirect lookup
Computing the `request.path` property in `webob/request.py` involves
reading values from the environment, then decoding, URL-quoting and
concatenating the results. This has enough overhead that it is worth
doing once rather than for each of the 214 redirect rules that we have,
on each request to h. Thread profiling via New Relic suggests this
accounts for ~8% of active request processing time.

A better but slightly more complex optimization would be to use a more
efficient data structure for matching the request path against the
redirect rules.