Remove New Relic badge metrics
`record_metrics` accounted for about 20% of the time spent in the badge
endpoint, most of this in resolving the user's authentication status and fetching
the corresponding `h.models.User` object from the DB.

Since we've now learned what we needed to about the results of badge
lookups, that the vast majority of requests return 0, I think we can
safely remove this. We can always get the code back from the history if
we need it again in future.