Add test for `DeletedFilter`
Unlike the rest of the fields which are set by AnnotationSearchIndexPresenter
when the annotation is indexed, the `deleted` field is set in the app by
`h.search.index.delete` which replaces the annotation's entire body in
the index. To ensure the delete operation/deleted filter match up, I've re-used
`h.search.index.delete` in the tests.

Add an optional `refresh` kwarg to `delete` to force a refresh so that
the change is immediately visible in search. Otherwise there will be a
brief delay until searches reflect the deletion.