Avoid loading all flagged userids to get the NIPSA status for a single user
`NipsaService.is_flagged` is currently used when indexing an annotation
into Elasticsearch and when processing an annotation event in the
WebSocket. In both contexts it is only called with a single userid.

Therefore it is more efficient to look up just that user and return
their NIPSA status rather than fetch _all_ users with `nipsa=True` and
check for the ID in the corresponding set of userids.