Remove old tests for `h.search.core`
The original tests which did not use Elasticsearch have been replaced
with more integrated tests which do. Having tests which actually use
Elasticsearch makes it possible to use them to assist with upgrading ES
versions and also provides more confidence that if the tests pass,
search functionality will actually work as expected in the app.

In some cases it may still make sense to mock ES for specific tests, eg.
to test specific edge cases or error conditions, or if the logic to
generate the ES query params is particularly complex.