Adapt filter query syntax depending on Elasticsearch version
The syntax for executing queries in a _filter context_ changed in
Elasticsearch 2.x. See [1]

Two pure unit tests have been added for the `Builder` class which check
the JSON output of `build(params)`. Once the tests which actually
execute queries against ES are working with ES 6, these will be
obsoleted.

[1] https://www.elastic.co/guide/en/elasticsearch/reference/6.2/query-dsl-filtered-query.html