Re-add NIPSA flag cache
The previous cache slowed down indexing and WebSocket message processing
tasks, but was valuable in one context: when reindexing all annotations
or all annotations for a specific user. In that context it avoided the
need to query for a user by userid for each annotation being indexed in
order to retrieve their nipsa status.

This commit re-introduces the cache but _only_ populates it when
`flagged_userids` is invoked. This means that requests which only query
the NIPSA status for one or a small number of users will not pay the
price of populating it.