Rename `NipsaService.flagged_userids`
Change `flagged_userids` from a property to a method and add a
`fetch_all_` prefix to make it more explicit that this code does a
potentially expensive database query, depending on the number of users
in the system. This is also more consistent with the API of eg. `UserService`.

The property => method change will also make it a little easier to mock
in an upcoming PR.